Alert

+++ html/Scripts/ansatte.js
@@ -37,9 +37,11 @@ AnsatteModul.getAllEmployes().forEach((ansatt, mIndex) => {
     const removeItem = () => {
         document.querySelectorAll("#delete-btn").forEach (button => { // identifies the delete button
             button.addEventListener("click", ( e ) => { // identifies the delete button was clicked
+                if (confirm("Hei! Er du helt sikkert du vil slette?")) {
                 const employeIndex = button.value; // finds the value (array index) that we added to the button HTML above
                 AnsatteModul.deleteItem(employeIndex); // passes that index into the deleteItem function within the Module
                 printEmployes(); //runs printMeny again to update HTML on the page to show item was deleted
+                }
             })
         })
